Organize components into folders
