+++ src/Components/Pages/Glossary.js
@@ -15,9 +15,9 @@ import IconButton from "@mui/material/IconButton";
 import UndoOutlinedIcon from "@mui/icons-material/UndoOutlined";
 import Box from "@mui/material/Box";
 
-import { updateRequest, createKey } from "../utils";
-import { unknownSectionName, UNKNOWN_TAG } from "../constants";
-import DepartmentFormControl from "./DepartmentFormControl";
+import { updateRequest, createKey } from "../../utils";
+import { unknownSectionName, UNKNOWN_TAG } from "../../constants";
+import DepartmentFormControl from "../Utils/DepartmentFormControl";
 
 function Glossary(props) {
   const {
